eliminar-ajax todos menos role

+++ app/Http/Controllers/UserController.php
@@ -5,7 +5,8 @@ namespace App\Http\Controllers;
 use App\User;
 use Caffeinated\Shinobi\Models\Role;
 use Illuminate\Http\Request;
-
+use App\Http\Controllers\HomeController as home;
+use Auth;
 
 class UserController extends Controller
 {
@@ -15,7 +16,7 @@ class UserController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
+    {   
         $users=User::paginate(10);
 
         return view("users.index",compact("users"));
@@ -51,9 +52,8 @@ class UserController extends Controller
     //public function show($id)
     public function show(User $user)
     {
-        
+        //dd($user);
         //$user=User::find($id);//->where("user","[0-9]+");
-
         return view("users.show",compact("user"));
     }
 
@@ -65,6 +65,7 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        
         //$user=User::find($id);
         $roles=Role::get();
 
@@ -91,9 +92,19 @@ class UserController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(User $user)
-    {
-        $user->delete();
-        return back();
+    public function destroy(Request $request)
+    {        
+        if($request->ajax()){
+            $div=home::ruta();
+            $user=User::where("id",$request->id)->first();
+            $user->delete();
+            $users=User::paginate(10);
+            $dato=view("users.table_users",compact("users"))->render();
+            return response()->json(["dato"=>$dato,"div"=>$div]);
+
+        }
+        
+        
+        
     }
 }
